# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: Beorimi API
  version: 0.4.0
components:
  securitySchemes:
    supabaseJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /api/health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: Backend health
          content:
            application/json:
              schema:
                type: object
                required: [status, service]
                properties:
                  status:
                    type: string
                  service:
                    type: string
  /api/uploads:
    post:
      operationId: createUploadUrl
      security:
        - supabaseJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/upload-url-request.json
      responses:
        "200":
          description: One-time Supabase Storage upload token scoped to the authenticated user and upload size limit
          content:
            application/json:
              schema:
                $ref: ../schemas/upload-url-response.json
        "400":
          description: Invalid filename, media type, or file size
        "401":
          description: Missing or invalid Supabase token
        "403":
          description: Current privacy consent has not been recorded
        "503":
          description: Upload storage is not configured
  /api/analyses:
    post:
      operationId: createAnalysis
      security:
        - supabaseJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/analysis-job-request.json
      responses:
        "202":
          description: Analysis was queued
          content:
            application/json:
              schema:
                $ref: ../schemas/analysis-job.json
        "400":
          description: Image key does not belong to the authenticated user
        "401":
          description: Missing or invalid Supabase token
        "403":
          description: Current privacy consent has not been recorded
        "404":
          description: Uploaded image does not exist
        "429":
          description: Authenticated account has used all five MVP analyses
        "503":
          description: Analysis worker is not configured
  /api/account:
    delete:
      operationId: deleteCurrentAccount
      security:
        - supabaseJwt: []
      responses:
        "204":
          description: Authenticated account, server analysis records, and uploaded images were deleted
        "401":
          description: Missing or invalid Supabase token
        "503":
          description: Account deletion could not be completed
  /api/analyses/{analysis_id}:
    get:
      operationId: getAnalysis
      security:
        - supabaseJwt: []
      parameters:
        - in: path
          name: analysis_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Analysis owned by the authenticated user
          content:
            application/json:
              schema:
                $ref: ../schemas/analysis-job.json
        "401":
          description: Missing or invalid Supabase token
        "404":
          description: Analysis does not exist or belongs to another user